Refactored project

+++ src/main/java/io/mountblue/blogapplication/service/impl/PostServiceImpl.java
@@ -6,6 +6,7 @@ import io.mountblue.blogapplication.entity.Post;
 import io.mountblue.blogapplication.entity.Tag;
 import io.mountblue.blogapplication.entity.User;
 import io.mountblue.blogapplication.exception.ResourceNotFoundException;
+import io.mountblue.blogapplication.exception.UserNotAuthenticatedException;
 import io.mountblue.blogapplication.repository.CommentRepository;
 import io.mountblue.blogapplication.repository.PostRepository;
 import io.mountblue.blogapplication.repository.TagRepository;
@@ -22,48 +23,53 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @Transactional
 @Service("postService")
 public class PostServiceImpl implements PostService {
     @Autowired
-    private PostRepository postRepository;
-
-    @Autowired
-    private TagRepository tagRepository;
-
-    @Autowired
     UserRepository userRepository;
 
     @Autowired
     CommentRepository commentRepository;
+
+    @Autowired
+    private PostRepository postRepository;
+
+    @Autowired
+    private TagRepository tagRepository;
     @Autowired
     private ModelMapper modelMapper;
 
     @Override
-    public PostDTO findPostById(Long id) {
+    public PostDetailsDTO findPostById(Long id) {
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Post with ID " + id + " not found"));
 
-        PostDTO postDTO = modelMapper.map(post, PostDTO.class);
+        PostDetailsDTO postDTO = modelMapper.map(post, PostDetailsDTO.class);
         return postDTO;
     }
 
     @Override
-    public PostDTO savePost(PostDTO postDTO) {
+    public PostDetailsDTO savePost(PostRequestDTO postDTO) {
         Post post = modelMapper.map(postDTO, Post.class);
         User user = getAuthenticatedUser()
-                .orElseThrow(() -> new IllegalStateException("User is not logged in."));
+                .orElseThrow(() -> new UserNotAuthenticatedException("User is not logged in."));
 
         post.setAuthor(user);
         post.setCreatedAt(LocalDateTime.now());
         post.setUpdatedAt(LocalDateTime.now());
         post.setPublishedAt(LocalDateTime.now());
-        post.setPublished(true);
+
+        if (post.isPublished()) {
+            post.setPublishedAt(LocalDateTime.now());
+        }
 
         postDTO.getTagsList().forEach(tagName -> {
             Tag tag = tagRepository.findByName(tagName).orElseGet(() -> new Tag(tagName));
@@ -71,52 +77,53 @@ public class PostServiceImpl implements PostService {
         });
 
         Post savedPost = postRepository.save(post);
-        return modelMapper.map(savedPost, PostDTO.class);
+        return modelMapper.map(savedPost, PostDetailsDTO.class);
     }
 
     @Override
-    public PostDTO updatePost(Long id, PostDTO postDTO) {
-        Post post = postRepository.findById(id)
+    public PostDetailsDTO updatePost(Long id, PostRequestDTO postRequestDTO) {
+        Post existingPost = postRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Post with ID " + id + " not found"));
 
-        post.setTitle(postDTO.getTitle());
-        post.setExcerpt(postDTO.getExcerpt());
-        post.setContent(postDTO.getContent());
-        post.setUpdatedAt(LocalDateTime.now());
+        existingPost.setTitle(postRequestDTO.getTitle());
+        existingPost.setExcerpt(postRequestDTO.getExcerpt());
+        existingPost.setContent(postRequestDTO.getContent());
+        existingPost.setUpdatedAt(LocalDateTime.now());
 
-        post.getTags().clear();
-        postDTO.getTagsList().forEach(tagName -> {
+        existingPost.getTags().clear();
+        postRequestDTO.getTagsList().forEach(tagName -> {
             Tag tag = tagRepository.findByName(tagName).orElseGet(() -> new Tag(tagName));
-            post.addTag(tag);
+            existingPost.addTag(tag);
         });
 
-        Post updatedPost = postRepository.save(post);
-        return modelMapper.map(updatedPost, PostDTO.class);
+        Post updatedPost = postRepository.save(existingPost);
+        return modelMapper.map(updatedPost, PostDetailsDTO.class);
     }
 
     @Override
-    public List<PostSummaryDTO> deletePostById(Long id) {
+    public void deletePostById(Long id) {
+        if (!postRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Post with id " + id + " not found");
+        }
+
         postRepository.deleteById(id);
-        return postRepository.findAll()
-                .stream()
-                .map(post -> modelMapper.map(post, PostSummaryDTO.class))
-                .collect(Collectors.toList());
     }
 
+
     @Override
-    public PostDTO addComment(CommentDTO commentDTO, Long postId) {
+    public PostDetailsDTO addComment(CommentRequestDTO commentRequestDTO, Long postId) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new ResourceNotFoundException("Post with ID " + postId + " not found"));
 
-        Comment comment = modelMapper.map(commentDTO, Comment.class);
+        Comment comment = modelMapper.map(commentRequestDTO, Comment.class);
 
         Optional<User> user = getAuthenticatedUser();
 
-        if(user.isPresent()){
+        if (user.isPresent()) {
             comment.setAuthor(user.get());
             comment.setEmail(user.get().getEmail());
             comment.setName(user.get().getFullName());
-        }else{
+        } else {
             User guest = userRepository.findById(8L).orElseThrow();
             comment.setAuthor(guest);
         }
@@ -124,32 +131,20 @@ public class PostServiceImpl implements PostService {
         comment.setCreatedAt(LocalDateTime.now());
         comment.setUpdatedAt(LocalDateTime.now());
         comment.setPost(post);
+        commentRepository.save(comment);
 
         post.addComment(comment);
         Post savedPost = postRepository.save(post);
 
-        return modelMapper.map(savedPost, PostDTO.class);
+        return modelMapper.map(savedPost, PostDetailsDTO.class);
     }
 
     @Override
-    public PostDTO deleteComment(Long postId, Long commentId) {
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new ResourceNotFoundException("Post with ID " + postId + " not found"));
-
-        Comment comment = post.getComments().stream()
-                .filter(c -> c.getId().equals(commentId))
-                .findFirst()
-                .orElseThrow(() -> new ResourceNotFoundException("Comment with ID " + commentId + " not found"));
-
-        post.removeComment(comment);
-        Post updatedPost = postRepository.save(post);
-
-        return modelMapper.map(updatedPost, PostDTO.class);
-    }
-
-    @Override
-    public List<String> findAllTags() {
-        return tagRepository.findAll().stream().map(Tag::getName).collect(Collectors.toList());
+    public Set<TagDTO> findAllTags() {
+        List<Tag> tags = tagRepository.findAll();
+        return tags.stream()
+                .map(tag -> modelMapper.map(tag, TagDTO.class))
+                .collect(Collectors.toSet());
     }
 
     @Override
@@ -161,35 +156,34 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public Page<PostSummaryDTO> findAllPosts(PostFilterDTO filterDTO) {
-        Pageable pageable = PageRequest.of(filterDTO.getPageNumber(), filterDTO.getPageSize(), Sort.by(filterDTO.getOrder().equals("latest") ? Sort.Direction.DESC : Sort.Direction.ASC, "createdAt"));
+    public Page<PostSummaryDTO> findAllPosts(int pageNumber, int pageSize, String order) {
+        Pageable pageable = PageRequest.of(pageNumber, pageSize, Sort.by(order.equalsIgnoreCase("latest") ? Sort.Direction.DESC : Sort.Direction.ASC, "createdAt"));
         Page<Post> paginatedPosts = postRepository.findAll(pageable);
         return paginatedPosts.map(post -> modelMapper.map(post, PostSummaryDTO.class));
     }
 
     @Override
-    public Page<PostSummaryDTO> searchQueryPosts(PostFilterDTO filterDTO) {
-        Pageable pageable = PageRequest.of(filterDTO.getPageNumber(), filterDTO.getPageSize(), Sort.by(filterDTO.getOrder().equals("latest") ? Sort.Direction.DESC : Sort.Direction.ASC, "createdAt"));
-        Page<Post> paginatedPosts = postRepository.searchPosts(filterDTO.getSearchQuery(), pageable);
+    public Page<PostSummaryDTO> searchQueryPosts(int pageNumber, int pageSize, String searchQuery, String order) {
+        Pageable pageable = PageRequest.of(pageNumber, pageSize, Sort.by(order.equalsIgnoreCase("latest") ? Sort.Direction.DESC : Sort.Direction.ASC, "createdAt"));
+        Page<Post> paginatedPosts = postRepository.searchPosts(searchQuery, pageable);
         return paginatedPosts.map(post -> modelMapper.map(post, PostSummaryDTO.class));
     }
 
     @Override
-    public Page<PostSummaryDTO> findFilteredPosts(PostFilterDTO filterDTO) {
-        Pageable pageable = PageRequest.of(filterDTO.getPageNumber(),
-                filterDTO.getPageSize(),
-                Sort.by(filterDTO.getOrder().equals("latest") ? Sort.Direction.DESC : Sort.Direction.ASC, "createdAt"));
+    public Page<PostSummaryDTO> findFilteredPosts(int pageNumber, int pageSize, List<Long> tags, List<Long> authors, LocalDate startDate, LocalDate endDate, String order) {
+        Pageable pageable = PageRequest.of(pageNumber, pageSize,
+                Sort.by(order.equalsIgnoreCase("latest") ? Sort.Direction.DESC : Sort.Direction.ASC, "createdAt"));
 
         Page<Post> paginatedPosts;
 
-        if (filterDTO.getDate() == null && filterDTO.getAuthors() == null && filterDTO.getTags() == null) {
+        if (tags == null && authors == null && startDate == null && endDate == null) {
             paginatedPosts = postRepository.findAll(pageable);
-        }else {
+        } else {
             paginatedPosts = postRepository.findByAuthorInOrTagsNameInOrCreatedAtBetween(
-                    filterDTO.getAuthors(),
-                    filterDTO.getTags(),
-                    filterDTO.getStartOfDay(),
-                    filterDTO.getEndOfDay(),
+                    tags,
+                    authors,
+                    startDate.atStartOfDay(),
+                    endDate.plusDays(1).atStartOfDay(),
                     pageable);
         }
 
@@ -206,22 +200,13 @@ public class PostServiceImpl implements PostService {
         return Optional.empty();
     }
 
-    public boolean isCreator(Long id){
+    public boolean isCreator(Long id) {
         User user = getAuthenticatedUser()
-                .orElseThrow(() -> new IllegalStateException("User is not logged in."));
+                .orElseThrow(() -> new UserNotAuthenticatedException("User is not logged in."));
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Post not found."));
 
         return user.getId() == post.getAuthor().getId();
     }
 
-    public boolean isOwnerOfCommentedPost(Long id){
-        User user = getAuthenticatedUser()
-                .orElseThrow(() -> new IllegalStateException("User is not logged in."));
-
-        Comment comment = commentRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Comment not found."));
-
-        return comment.getPost().getAuthor().getId() == user.getId();
-    }
 }
